Add Diagnose method to runner.Runner interface
