+++ internal/cmdrunner/cmd_runner.go
@@ -106,3 +106,17 @@ func (c *CmdRunner) Name() string {
 func (c *CmdRunner) ID() string {
 	return fmt.Sprintf("%d", c.pid)
 }
+
+// peTypes is a list of Portable Executable (PE) machine types from https://learn.microsoft.com/en-us/windows/win32/debug/pe-format
+// mapped to GOARCH types. It is not comprehensive, and only includes machine types that Go supports.
+var peTypes = map[uint16]string{
+	0x14c:  "386",
+	0x1c0:  "arm",
+	0x6264: "loong64",
+	0x8664: "amd64",
+	0xaa64: "arm64",
+}
+
+func (c *CmdRunner) Diagnose(_ context.Context) string {
+	return additionalNotesAboutCommand(c.cmd.Path)
+}
