+++ internal/cmdrunner/testdata/minimal.go
